refactor: Change redux directory structure
