--- client/src/redux/epics/index.ts
@@ -1,3 +1,0 @@
-import { combineEpics } from "redux-observable";
-const rootEpic = combineEpics();
-export default rootEpic;
