+++ client/src/redux/modules/editorReducer.ts
@@ -1,4 +1,4 @@
-import { Zettel } from "../models/Zettel";
+import { Zettel } from "../../models/Zettel";
 
 const SET_VIEWER = "editor/SET_VIEWER" as const;
 const START_EDIT = "editor/START_EDIT" as const;
